Revert "chore: harden backend app, secure auth, add tests & frontend resiliency"
