+++ backend/.env.example
@@ -13,5 +13,3 @@ TWILIO_FROM_NUMBER=
 REDIS_URL=
 SKIP_MONGO=false
 SKIP_FIREBASE=true
-MONGOMS_VERSION=7.0.14
-MONGOMS_DISTRO=ubuntu-22.04
