+++ backend/middleware/combinedAuth.js
@@ -1,5 +1,4 @@
 const jwt = require('jsonwebtoken');
-const User = require('../models/User');
 
 const extractToken = (req) =>
   req.cookies?.accessToken ||
@@ -69,11 +68,18 @@ exports.authenticate = async (req, res, next) => {
 };
 
 exports.authorize = (...roles) => (req, res, next) => {
-  if (!req.user) {
-    return res.status(401).json({ message: 'Unauthorized' });
-  }
-  if (roles.length && !roles.includes(req.user.role)) {
+  const currentRole = req.user?.role;
+  if (!currentRole || (roles.length && !roles.includes(currentRole))) {
     return res.status(403).json({ message: 'Forbidden' });
   }
   return next();
+};
+
+exports.authorize = (roles = []) => {
+  return (req, res, next) => {
+    if (!req.user || !roles.includes(req.user.role)) {
+      return res.status(403).json({ message: "Unauthorized Role" });
+    }
+    next();
+  };
 };
