--- backend/routes/logs.js
@@ -1,21 +1,0 @@
-const express = require('express');
-const Log = require('../models/Log');
-const { authenticate, authorize } = require('../middleware/combinedAuth');
-
-const router = express.Router();
-
-router.get('/', authenticate, authorize('superadmin'), async (req, res) => {
-  const logs = await Log.find()
-    .populate('tenant', 'name')
-    .populate('user', 'name')
-    .sort({ createdAt: -1 })
-    .limit(100);
-  res.json(logs);
-});
-
-router.delete('/', authenticate, authorize('superadmin'), async (req, res) => {
-  await Log.deleteMany({});
-  res.json({ message: 'Logs cleared.' });
-});
-
-module.exports = router;
