--- backend/routes/tenants.js
@@ -1,27 +1,0 @@
-const express = require('express');
-const Tenant = require('../models/Tenant');
-const { authenticate, authorize } = require('../middleware/combinedAuth');
-
-const router = express.Router();
-
-router.get('/', authenticate, authorize('superadmin'), async (req, res) => {
-  const tenants = await Tenant.find().sort({ createdAt: -1 });
-  res.json(tenants);
-});
-
-router.patch('/:id/status', authenticate, authorize('superadmin'), async (req, res) => {
-  const { isActive } = req.body;
-  const tenant = await Tenant.findByIdAndUpdate(
-    req.params.id,
-    { isActive: Boolean(isActive) },
-    { new: true }
-  );
-
-  if (!tenant) {
-    return res.status(404).json({ message: 'Tenant not found.' });
-  }
-
-  res.json(tenant);
-});
-
-module.exports = router;
