+++ frontend/src/pages/Dashboard.jsx
@@ -2,35 +2,14 @@ import { useEffect, useState } from "react";
 import { Helmet } from "react-helmet-async";
 import { useNavigate } from "react-router-dom";
 import api from "../api/axios";
-import { getApiBaseUrl } from "../api/baseUrl";
-import { safeStorage } from "../utils/storage";
 
 export default function Dashboard() {
   const navigate = useNavigate();
   const [bookings, setBookings] = useState([]);
-  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    api
-      .get("/booking/customer/bookings")
-      .then((res) => setBookings(res.data))
-      .catch((error) => console.error("Fetch error:", error))
-      .finally(() => setLoading(false));
+    api.get("/booking/customer/bookings").then((res) => setBookings(res.data));
   }, []);
-
-  const API_BASE = getApiBaseUrl();
-
-  const downloadBilty = (id) =>
-    window.open(
-      `${API_BASE}/booking/${id}/bilty?token=${safeStorage.get("accessToken")}`,
-      "_blank"
-    );
-
-  const downloadInvoice = (id) =>
-    window.open(
-      `${API_BASE}/booking/${id}/invoice?token=${safeStorage.get("accessToken")}`,
-      "_blank"
-    );
 
   return (
     <div className="page">
@@ -53,9 +32,7 @@ export default function Dashboard() {
 
         <section className="stack">
           <h3 className="section-title">Your Bookings</h3>
-          {loading ? (
-            <p className="text-muted">Loading bookings...</p>
-          ) : bookings.length === 0 ? (
+          {bookings.length === 0 ? (
             <div className="card">
               <p style={{ margin: 0, fontWeight: 600 }}>No bookings yet</p>
               <p className="text-muted" style={{ margin: "6px 0 0" }}>
@@ -81,20 +58,6 @@ export default function Dashboard() {
                 >
                   {b.status}
                 </span>
-                <div className="toolbar" style={{ marginTop: 12 }}>
-                  <button
-                    className="btn btn-secondary"
-                    onClick={() => downloadBilty(b._id)}
-                  >
-                    Bilty PDF
-                  </button>
-                  <button
-                    className="btn btn-secondary"
-                    onClick={() => downloadInvoice(b._id)}
-                  >
-                    Invoice PDF
-                  </button>
-                </div>
               </div>
             ))
           )}
