+++ frontend/src/pages/Login.jsx
@@ -46,8 +46,8 @@ export default function Login() {
       return "Please enter a valid email address.";
     }
 
-    if (password.length < 8) {
-      return "Password must be at least 8 characters.";
+    if (password.length < 6) {
+      return "Password must be at least 6 characters.";
     }
 
     return null;
