--- frontend/vitest.config.js
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config";
-import react from "@vitejs/plugin-react";
-
-export default defineConfig({
-  plugins: [react()],
-  test: {
-    environment: "jsdom",
-    setupFiles: "./src/test/setupTests.js",
-    globals: true,
-  },
-});
