+++ render.yaml
@@ -8,7 +8,7 @@ services:
     previewsEnabled: true
     buildCommand: npm ci
     startCommand: npm start
-    healthCheckPath: /api/health
+    healthCheckPath: /
     envVars:
       - key: NODE_ENV
         value: production
